Fix #16
The time sliders are not displayed on the admin product page auction tab
from the WooCommerce Simple Auctions plugin.

The plugin allows to add a new product type named Auction on the admin
edit product page. If the product type is Auction then it adds a new tab
in Product data meta box name Auction which allows to set the timings
and it has time sliders to set hours and minutes. Which is not displayed
when the plugin is active.