Version 2.2
Changed the version number 2.2

Change log is as below:

* The Delivery Date field label, field placeholder text and the field
note text in Order Delivery Date -> Appearance can now be translated
with WPML plugin.
* Until now, only admin user was able to access the Order Delivery Date
settings page. From this update, users with the role 'Shop manager' will
be able to access and edit the Order Delivery Date settings.
* The calendar was not appearing on the checkout page when clicked on
the Delivery Date field on the checkout page. This issue is specific to
the 'Stockholm' theme. This is fixed now.