Version 2.3
Changed the version number 2.2

Change log is as below:

* Calendar next and previous arrows were not displayed correctly with
the Galleria theme on
the checkout page. This is fixed now.
* Delivery Date was not displayed on the invoice and packing list due to
a hook being
deprecated from WooCommerce Print Invoices/Packing Lists plugin. This
issue is fixed now.
* A deprecated hook was used to add the Delivery Date value in customer
notification email. This is fixed now.
* Calendar weekdays were overlapping in the calendar on the checkout
page when the "Number of Months" was set to 2. This is fixed now.
* Delivery Dates were not translated on the Order Received page, My
Account page and WooCommerce -> Orders page when date language is set to
Dutch-Belgian. This is fixed now.