Delivery Date field can be enabled or disabled.
A checkbox named 'Enable Delivery Date:' is added in the Date Settings
tab.

When the checkbox is enabled, the Delivery Date field will be enabled to
capture the delivery date for an order on the checkout page otherwise
the field will not appear on the checkout page.

The default value of this setting is off. When the plugin is updated
from lower version, the value for this setting will be enabled or on.