Disable field on checkout page for some products.
From now, the Delivery Date field can be disabled on the checkout page
for Virtual and Featured products.

A setting named 'Disable the Delivery Date Field for' is added in the
Appearance tab. When a checkbox named 'Virtual Products' is enabled then
the delivery date field will be disabled for only virtual products.

When a checkbox 'Featured products' is enabled then the delivery date
field will be disabled for only Featured products.

The default value for both the checkboxes is disabled (off).