Version 2.6
= 2.6 =

* From now on the Delivery Date can be enabled or disabled for capturing
the delivery date. A checkbox named as 'Enable Delivery Date' is added
in the Date Settings tab.

The default value of this setting is disabled, which when enabled will
allow clients to capture the Delivery Date for an order on the checkout
page.
* The Delivery Date field can also be disabled for Virtual and Featured
products on the checkout page.

A setting named 'Disable the Delivery Date Field for' is added in the
Appearance tab. The default value for both the checkboxes is disabled.
When a checkbox named 'Virtual Products' is enabled then the delivery
date field will be disabled for only virtual products.

When a checkbox 'Featured Products' is enabled then the delivery date
field will be disabled for only featured products.
* A notice will be displayed when the Order Delivery Date for
WooCommerce plugin is activated but the WooCommerce plugin is
deactivated or it is not installed. The plugin will be deactivated
automatically.
* The width of the Delivery Date field will be the same as that of the
other fields on the checkout page.
* Calendar was not working on the checkout page when Impreza theme was
active.
* The cursor for the Delivery Date field will be text instead of a
pointer.