Disabled weekdays and holidays will not be calculated in Minimum Delivery Time
Currently minimum delivery time in hours is applied on disabled weekdays by default. A setting is added named as 'Apply Minimum Delivery Time for non working weekdays' under Date settings tab. The default value of this checkbox will be on if plugin is updated else it will be off if installed for the first. And if this setting is on then the minimum delivery time will be applied on the disabled weekdays else not.

For holidays, no setting is added, the minimum delivery time will be not applied on holidays by default.