Admin notices for pro version and for other plugins
We have added 5 notices at the admin side. It will display after 15 days, 22 days, 30 days and 37 days after the first notice has been ignored.

If The Order Delivery Date Pro for WooCommerce plugin is activated then none of the notices will be shown.

In Second and Fifth notice that is on 15th day and on 37th day, the notice of 'Post Delivery Product Reviews' addon will be displayed.

If the addon is active then the next notice which is AC lite notice will be displayed for the second notice.

In third notice that is 22 days, we are showing it for AC lite, so AC Pro is activated then we wil not display the same notice.

On dismissable of 3rd notice, again Orddd Pro notice will be shown.