Fix #193
Calendar is not working on the checkout page with the Prohauz Child theme due to custom code they have added to add defer onload in the js files while including them.

// Defer Javascripts
// Defer jQuery Parsing using the HTML5 defer property
if (!(is_admin() )) {
    function defer_parsing_of_js ( $url ) {
        if ( FALSE === strpos( $url, '.js' ) ) return $url;
        if ( strpos( $url, 'jquery.js' ) ) return $url;
        // return "$url' defer ";
        return "$url' defer onload='";
    }
    add_filter( 'clean_url', 'defer_parsing_of_js', 11, 1 );
}

The issue was due to the esc_url function we were using in wp_enqueue_scripts. And as we don't need to use esc_url in the wp_enqueue_scripts function we are removing it.